リファクタリング #13
ログアウト処理のリファクタリングとそれに伴うログイン処理の修正
