--- frontend/src/views/api/login.js
@@ -1,9 +1,0 @@
-import { apiClient } from "./client";
-
-export const login = (username, password) => {
-  return apiClient.post(
-    "login",
-    { username, password },
-    { withCredentials: true }
-  )
-};
