--- frontend/src/views/lib/baseUrl.js
@@ -1,3 +1,0 @@
-const appScheme = import.meta.env.VITE_APP_SCHEME;
-const backendHost = import.meta.env.VITE_BACKEND_HOST;
-export const backendUrl = appScheme + "://" + backendHost + "/";
